Small fixes required to make it work on x86_64 architecture.
1. The comparision of ULONG field against -1UL is invalid on x86_64,
since the -1UL equals 0xffffffffffffffff and does not fit into ULONG.

2. The oop.library used utility.library but never opened it. Changed.


git-svn-id: https://svn.aros.org/svn/aros/trunk@26484 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ rom/oop/support.c
@@ -123,7 +123,7 @@ BOOL init_methodbase(STRPTR interface_id, ULONG methodbase, ULONG *methodbase_pt
     idb = (struct iid_bucket *)iidtable->Lookup(iidtable, (IPTR)interface_id, (struct IntOOPBase *)OOPBase);
     if (idb)
     {
-	if (idb->methodbase == -1UL)
+	if (idb->methodbase == (ULONG)-1)
 	{
 	    idb->methodbase = methodbase;
 	}
